Moved component to election_overview/components
