+++ frontend/src/features/election_overview/components/AddFirstElection.stories.tsx
